Incorporate review comments
- Rename from Deinit_Required to Required_Deinit
- Adopt ASTRule protocol
