+++ CHANGELOG.md
@@ -32,7 +32,7 @@
 * Make `redundant_objc_attribute` rule autocorrectable.  
   [Daniel Metzing](https://github.com/dirtydanee)
   
-* Add `deinit_required` opt-in rule to ensure that all classes have a deinit
+* Add `required_deinit` opt-in rule to ensure that all classes have a deinit
   method. The purpose of this is to make memory leak debugging easier so all
   classes have a place to set a breakpoint to track deallocation.  
   [Ben Staveley-Taylor](https://github.com/BenStaveleyTaylor)
