+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -22,7 +22,6 @@ public let masterRuleList = RuleList(rules: [
     ConvenienceTypeRule.self,
     CustomRules.self,
     CyclomaticComplexityRule.self,
-    DeinitRequiredRule.self,
     DeploymentTargetRule.self,
     DiscardedNotificationCenterObserverRule.self,
     DiscouragedDirectInitRule.self,
@@ -123,6 +122,7 @@ public let masterRuleList = RuleList(rules: [
     RedundantStringEnumValueRule.self,
     RedundantTypeAnnotationRule.self,
     RedundantVoidReturnRule.self,
+    RequiredDeinitRule.self,
     RequiredEnumCaseRule.self,
     ReturnArrowWhitespaceRule.self,
     ShorthandOperatorRule.self,
